@@ -1,18 +1,34 @@
 module github.com/mehdy/keepalived-exporter
 
-go 1.15
+go 1.19
+
+require (
+	github.com/docker/docker v20.10.23+incompatible
+	github.com/hashicorp/go-version v1.6.0
+	github.com/prometheus/client_golang v1.14.0
+	github.com/sirupsen/logrus v1.9.0
+)
 
 require (
 	github.com/Microsoft/go-winio v0.4.15 // indirect
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/cespare/xxhash/v2 v2.1.2 // indirect
 	github.com/docker/distribution v2.7.1+incompatible // indirect
-	github.com/docker/docker v1.13.1
 	github.com/docker/go-connections v0.4.0 // indirect
 	github.com/docker/go-units v0.4.0 // indirect
-	github.com/hashicorp/go-version v1.2.1
-	github.com/magefile/mage v1.11.0 // indirect
+	github.com/gogo/protobuf v1.1.1 // indirect
+	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/matttproud/golang_protobuf_extensions v1.0.1 // indirect
+	github.com/moby/term v0.0.0-20221205130635-1aeaba878587 // indirect
+	github.com/morikuni/aec v1.0.0 // indirect
 	github.com/opencontainers/go-digest v1.0.0 // indirect
-	github.com/prometheus/client_golang v1.9.0
-	github.com/sirupsen/logrus v1.8.0
-	golang.org/x/net v0.0.0-20201202161906-c7110b5ffcbb // indirect
-	golang.org/x/sys v0.0.0-20210304152209-afaa3650a925 // indirect
+	github.com/opencontainers/image-spec v1.0.2 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/prometheus/client_model v0.3.0 // indirect
+	github.com/prometheus/common v0.37.0 // indirect
+	github.com/prometheus/procfs v0.8.0 // indirect
+	golang.org/x/net v0.0.0-20220225172249-27dd8689420f // indirect
+	golang.org/x/sys v0.4.0 // indirect
+	google.golang.org/protobuf v1.28.1 // indirect
+	gotest.tools/v3 v3.4.0 // indirect
 )
